HBASE-9248 Place holders for tags in 0.96 to accommodate tags in 0.98
git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.95@1516615 13f79535-47bb-0310-9956-ffa450edef68

+++ hbase-common/src/main/java/org/apache/hadoop/hbase/KeyValue.java
@@ -2880,7 +2880,7 @@ public class KeyValue implements Cell, HeapSize, Cloneable {
   }
 
   @Override
-  public int getTagsLength() {
+  public short getTagsLength() {
     throw new UnsupportedOperationException("Not implememnted");
   }
